Conflicts:
	opencv/face.py

+++ opencv/face.py
@@ -58,7 +58,9 @@ def main():
 def create_image_section(inputfile, ext, host, fileid, key):
     global logger, sslVerify
     logger.debug("INSIDE: create_image_section")
-    (fd, sectionfile)=tempfile.mkstemp(suffix='.' + ext)
+    
+    sectionfile=None
+    
     try:
         #extract face from images using opencv face detector
         #face_cascade = cv2.CascadeClassifier('/opt/local/share/OpenCV/haarcascades/haarcascade_frontalface_alt.xml')
@@ -80,7 +82,8 @@ def create_image_section(inputfile, ext, host, fileid, key):
             #create a preview for the section and upload the preview and its metadata
             for (x,y,w,h) in faces:
                 roi_color = img[y:y+h, x:x+w]
-                
+                (fd, sectionfile)=tempfile.mkstemp(suffix='.' + ext)
+                os.close(fd)
                       
                 cv2.imwrite(sectionfile, roi_color)
                 
@@ -105,7 +108,10 @@ def create_image_section(inputfile, ext, host, fileid, key):
                 logger.debug(("section id [%s]",sectionid))
 
                 url=host + 'api/previews?key=' + key
-                rc = requests.post(url, files={"File" : open(sectionfile, 'rb')}, verify=sslVerify)
+                f = open(sectionfile, 'rb')
+                files={"File" : f}
+                rc = requests.post(url, files=files, verify=sslVerify)
+                f.close()
                 rc.raise_for_status()
                 previewid = rc.json()['id']
                 logger.debug("preview id=[%s]",rc.json()['id'])
@@ -145,7 +151,7 @@ def create_image_section(inputfile, ext, host, fileid, key):
                 logger.debug("[%s] created section and previews of type %s", fileid, ext)
     finally:
         #os.remove(previewfile)
-        if os.path.isfile(sectionfile):     
+        if sectionfile is not None and os.path.isfile(sectionfile):     
             os.remove(sectionfile)  
 def get_image_data(imagefile):
     global logger
@@ -245,13 +251,12 @@ def on_message(channel, method, header, body):
                             properties=pika.BasicProperties(correlation_id = \
                                                         header.correlation_id),
                             body=json.dumps(statusreport))
-        if inputfile is not None:
+        if inputfile is not None and os.path.isfile(inputfile):
             try:
                 os.remove(inputfile)
-            except OSError:
-                pass
-            except UnboundLocalError:
-                pass    
+            except OSError as oserror:
+                logger.exception("[%s] error removing input file: \n %s", fileid, oserror)
+
 
 
 if __name__ == "__main__":
